# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RepoPilot",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/agent/sessions": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "List Agent Sessions",
        "operationId": "list_agent_sessions_api_v1_agent_sessions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_list_AgentSessionOut__"
                }
              }
            }
          }
        },
        "security": [
          {
            "OptionalAuthBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Create Agent Session",
        "operationId": "create_agent_session_api_v1_agent_sessions_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_AgentSessionOut_"
                }
              }
            }
          }
        },
        "security": [
          {
            "OptionalAuthBearer": []
          }
        ]
      }
    },
    "/api/v1/agent/sessions/{session_id}": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Get Agent Session",
        "operationId": "get_agent_session_api_v1_agent_sessions__session_id__get",
        "security": [
          {
            "OptionalAuthBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_AgentSessionDetailOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "agent"
        ],
        "summary": "Delete Agent Session",
        "operationId": "delete_agent_session_api_v1_agent_sessions__session_id__delete",
        "security": [
          {
            "OptionalAuthBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "agent"
        ],
        "summary": "Patch Agent Session",
        "operationId": "patch_agent_session_api_v1_agent_sessions__session_id__patch",
        "security": [
          {
            "OptionalAuthBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionUpdateBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_AgentSessionOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/sessions/{session_id}/chat": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Chat In Session",
        "operationId": "chat_in_session_api_v1_agent_sessions__session_id__chat_post",
        "security": [
          {
            "OptionalAuthBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentChatBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/chat": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Chat Legacy",
        "operationId": "chat_legacy_api_v1_agent_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OptionalAuthBearer": []
          }
        ]
      }
    },
    "/api/v1/agent/question": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Answer Question",
        "operationId": "answer_question_api_v1_agent_question_post",
        "security": [
          {
            "OptionalAuthBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "会话 ID（也可放 body）",
              "title": "Session Id"
            },
            "description": "会话 ID（也可放 body）"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentQuestionAnswer"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/analyze/{project_id}": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Analyze Project",
        "operationId": "analyze_project_api_v1_agent_analyze__project_id__post",
        "security": [
          {
            "OptionalAuthBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AnalyzeBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/import-assist": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Import Assist",
        "operationId": "import_assist_api_v1_agent_import_assist_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportAssistBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OptionalAuthBearer": []
          }
        ]
      }
    },
    "/api/v1/agent/graph-guide": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Graph Guide",
        "operationId": "graph_guide_api_v1_agent_graph_guide_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphGuideBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OptionalAuthBearer": []
          }
        ]
      }
    },
    "/api/v1/agent/trending-scout": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Trending Scout",
        "operationId": "trending_scout_api_v1_agent_trending_scout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrendingScoutBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OptionalAuthBearer": []
          }
        ]
      }
    },
    "/api/v1/agent/classify": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Classify Project",
        "operationId": "classify_project_api_v1_agent_classify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClassifyBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OptionalAuthBearer": []
          }
        ]
      }
    },
    "/api/v1/agent/note/generate": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Generate Note",
        "operationId": "generate_note_api_v1_agent_note_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteGenerateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OptionalAuthBearer": []
          }
        ]
      }
    },
...
    "/api/v1/agent/permissions": {
      "get": {
        "tags": [
          "agent"
        ],
        "summary": "Get Permissions",
        "operationId": "get_permissions_api_v1_agent_permissions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_AgentPermissionsOut_"
                }
              }
            }
          }
        },
        "security": [
          {
            "OptionalAuthBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "agent"
        ],
        "summary": "Patch Permissions",
        "description": "更新当前用户 Agent 工具权限（敏感能力开关）。",
        "operationId": "patch_permissions_api_v1_agent_permissions_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentPermissionsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_AgentPermissionsOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OptionalAuthBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentChatBody": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Message"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "AgentChatBody"
      },
...
      "DataResponse_AgentSessionDetailOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AgentSessionDetailOut"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DataResponse[AgentSessionDetailOut]"
      },
      "DataResponse_AgentSessionOut_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AgentSessionOut"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DataResponse[AgentSessionOut]"
      },
...
      "DataResponse_dict_": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DataResponse[dict]"
      },
...
      "DataResponse_list_AgentSessionOut__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/AgentSessionOut"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "DataResponse[list[AgentSessionOut]]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "OptionalAuthBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}